// Toolkit.Frontend/src/generated/Models/Core/TKJobsGetPagedHistoryRequestModel.ts
//     This code was generated by a Reinforced.Typings tool. 
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

export interface TKJobsGetPagedHistoryRequestModel
{
	SourceId: string;
	JobId: string;
	PageIndex: number;
	PageSize: number;
}