// Toolkit.Frontend/src/generated/Models/Core/TKJobDefinitionWithSourceViewModel.ts
//     This code was generated by a Reinforced.Typings tool. 
//     Changes to this file may cause incorrect behavior and will be lost if
//     the code is regenerated.

import { TKJobDefinitionViewModel } from './TKJobDefinitionViewModel';

export interface TKJobDefinitionWithSourceViewModel
{
	SourceId: string;
	Definition: TKJobDefinitionViewModel;
}